Вынести методы toString и formString в отдельный класс, сделать статическими. Вынести отдельный класс ManagerSaveException.
