+++ src/ru/yandex/javacource/sevrin/schedule/manager/FileBackedTaskManager.java
@@ -1,6 +1,7 @@
 package ru.yandex.javacource.sevrin.schedule.manager;
 
 import ru.yandex.javacource.sevrin.schedule.task.*;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -10,10 +11,8 @@ import java.util.Scanner;
 public class FileBackedTaskManager extends InMemoryTaskManager {
     File file;
 
-    public static class ManagerSaveException extends RuntimeException {
-        public ManagerSaveException(String message, Throwable cause) {
-            super(message, cause);
-        }
+    public FileBackedTaskManager(File file) {
+        this.file = file;
     }
 
     @Override
@@ -37,65 +36,6 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
         return id;
     }
 
-    public FileBackedTaskManager(File file) {
-        this.file = file;
-    }
-
-    public Task fromString(String str) {
-        try {
-            if (str == null || str.isEmpty()) {
-                throw new NullPointerException("Передано значение null!");
-            }
-            String[] parts = str.split(",");
-
-            if (parts.length < 5) {
-                throw new IllegalArgumentException("Неверный формат строки!");
-            }
-
-            int id = Integer.parseInt(parts[0]);
-            String type = parts[1].toUpperCase().trim();
-            String title = parts[2].trim();
-            Status status = Status.valueOf(parts[3].trim());
-            String description = parts[4].trim();
-
-            switch (type) {
-                case "TASK":
-                    return new Task(id, title, description, status);
-                case "EPIC":
-                    return new Epic(id, title, description, status);
-                case "SUBTASK":
-                    int epicId = Integer.parseInt(parts[5]);
-                    return new Subtask(id, title, description, status, epicId);
-                default:
-                    return null;
-            }
-        } catch (NullPointerException | IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-            return null;
-        }
-    }
-
-    public String toString(Task task) {
-        if (task instanceof Subtask subtask) {
-            return String.format("%d, %s, %s, %s, %s, %s",
-                    task.getId(),
-                    task.getClass().getSimpleName(),
-                    task.getTitle(),
-                    task.getStatus(),
-                    task.getDescription(),
-                    subtask.getEpicId()
-            );
-        } else {
-            return String.format("%d, %s, %s, %s, %s",
-                    task.getId(),
-                    task.getClass().getSimpleName(),
-                    task.getTitle(),
-                    task.getStatus(),
-                    task.getDescription()
-            );
-        }
-    }
-
     public static FileBackedTaskManager loadFromFile(File file) {
         FileBackedTaskManager manager = new FileBackedTaskManager(file);
 
@@ -110,13 +50,12 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
                     continue;
                 }
 
-                Task task = manager.fromString(line);
+                Task task = Formatter.fromString(line);
                 if (task != null) {
                     if (task instanceof Epic) {
                         manager.epics.put(task.getId(), (Epic) task);
                     } else if (task instanceof Subtask) {
                         manager.subtasks.put(task.getId(), (Subtask) task);
-                        // Добавляем подзадачу в эпик
                         Subtask subtask = (Subtask) task;
                         Epic epic = manager.epics.get(subtask.getEpicId());
                         if (epic != null) {
@@ -142,13 +81,13 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
         try (FileWriter writer = new FileWriter(file)) {
             writer.write("id,type,name,status,description,epic\n");
             for (Task task : getTasks()) {
-                writer.write(toString(task) + "\n");
+                writer.write(Formatter.toString(task) + "\n");
             }
             for (Epic epic : getEpics()) {
-                writer.write(toString(epic) + "\n");
+                writer.write(Formatter.toString(epic) + "\n");
             }
             for (Subtask subtask : getSubtasks()) {
-                writer.write(toString(subtask) + "\n");
+                writer.write(Formatter.toString(subtask) + "\n");
             }
 
         } catch (IOException e) {
