+++ src/ru/yandex/javacource/sevrin/schedule/manager/ManagerSaveException.java
@@ -1,0 +1,7 @@
+package ru.yandex.javacource.sevrin.schedule.manager;
+
+public class ManagerSaveException extends RuntimeException {
+    public ManagerSaveException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
