+++ src/ru/yandex/javacource/sevrin/schedule/task/Epic.java
@@ -12,6 +12,7 @@ public class Epic extends Task {
     public Epic(int id, String title, String description, Status status) {
         super(id, title, description, status);
     }
+
     public ArrayList<Integer> getSubtaskIds() {
         return subtaskIds;
     }
